+++ src/views/app-view.tsx
@@ -3,31 +3,45 @@
  *
  * SPDX-License-Identifier: AGPL-3.0-only
  */
-import React, { FC, Suspense, lazy, useEffect, useMemo, useState } from 'react';
+import React, { FC, Suspense, lazy, useEffect, useMemo, useState, useRef } from 'react';
 
-import { Container } from '@zextras/carbonio-design-system';
-import { FOLDERS, Spinner, setAppContext, useUserSettings } from '@zextras/carbonio-shell-ui';
+import {
+	FOLDERS,
+	Spinner,
+	setAppContext,
+	useUserSettings,
+	useLocalStorage
+} from '@zextras/carbonio-shell-ui';
 import { includes } from 'lodash';
 import moment from 'moment';
-import { Redirect, Route, Switch, useRouteMatch } from 'react-router-dom';
 
+import { FolderView, MailsListLayout } from './folder-view';
+import { LayoutSelector } from './layout-selector';
+import { LOCAL_STORAGE_LAYOUT, MAILS_VIEW_LAYOUTS } from '../constants';
 import { getFolderIdParts } from '../helpers/folders';
 import { useAppSelector } from '../hooks/redux';
 import { selectCurrentFolder } from '../store/conversations-slice';
-
-const LazyFolderView = lazy(
-	() => import(/* webpackChunkName: "folder-panel-view" */ './app/folder-panel')
-);
 
 const LazyDetailPanel = lazy(
 	() => import(/* webpackChunkName: "folder-panel-view" */ './app/detail-panel')
 );
 
+const DetailPanel = (): React.JSX.Element => (
+	<Suspense fallback={<Spinner />}>
+		<LazyDetailPanel />
+	</Suspense>
+);
+
 const AppView: FC = () => {
-	const { path } = useRouteMatch();
 	const [count, setCount] = useState(0);
 	const { zimbraPrefGroupMailBy, zimbraPrefLocale } = useUserSettings().prefs;
 	const currentFolderId = useAppSelector(selectCurrentFolder);
+	const containerRef = useRef<HTMLDivElement>(null);
+
+	const [listLayout] = useLocalStorage<MailsListLayout>(
+		LOCAL_STORAGE_LAYOUT,
+		MAILS_VIEW_LAYOUTS.LEFT_TO_RIGHT
+	);
 
 	const isMessageView = useMemo(
 		() =>
@@ -45,21 +59,12 @@ const AppView: FC = () => {
 	}, [count, isMessageView]);
 
 	return (
-		<Container orientation="horizontal" mainAlignment="flex-start">
-			<Container width="40%">
-				<Switch>
-					<Route path={`${path}/folder/:folderId/:type?/:itemId?`}>
-						<Suspense fallback={<Spinner />}>
-							<LazyFolderView />
-						</Suspense>
-					</Route>
-					<Redirect strict from={path} to={`${path}/folder/2`} />
-				</Switch>
-			</Container>
-			<Suspense fallback={<Spinner />}>
-				<LazyDetailPanel />
-			</Suspense>
-		</Container>
+		<LayoutSelector
+			listLayout={listLayout}
+			folderView={<FolderView listLayout={listLayout} containerRef={containerRef} />}
+			detailPanel={<DetailPanel />}
+			containerRef={containerRef}
+		/>
 	);
 };